Replace temporary file 2 with the original SinglePageReader
