+++ src/components/Reader/index.js
@@ -7,7 +7,6 @@ import compact from "lodash/compact";
 import type { Manga } from "@tachiweb/api-client";
 import type { ChapterType } from "types";
 import SinglePageReader from "components/Reader/SinglePageReader";
-import SinglePageReader2 from "components/Reader/SinglePageReader2";
 import WebtoonReader from "components/Reader/WebtoonReader";
 import { Helmet } from "react-helmet";
 import { chapterNumPrettyPrint } from "components/utils";
@@ -100,7 +99,7 @@ const Reader = ({ match: { params } }: Props) => {
         //   nextChapterUrl={nextChapterUrl()}
         //   prevChapterUrl={prevChapterUrl()}
         // />
-        <SinglePageReader2
+        <SinglePageReader
           mangaInfo={mangaInfo}
           chapter={chapter}
           pageCount={pageCount}
